backend/pyproject.toml
[project]
name = "aethera-backend"
version = "0.1.0"
description = "Core backend, geospatial engine, emissions, and orchestration for AETHERA."
authors = [{name = "AETHERA Team"}]
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "loguru>=0.7",
    "geopandas>=0.14",
    "shapely>=2.0",
    "pyogrio>=0.8",
    "rasterio>=1.3",
    "fiona>=1.9",
    "pandas>=2.2",
    "numpy>=1.26",
    "pyarrow>=15.0",
    "scikit-learn>=1.4",
    "xarray>=2024.1",
    "rioxarray>=0.15",
    "pyproj>=3.6",
    "requests>=2.32",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "psycopg[binary,pool]>=3.1",
    "celery>=5.4",
    "redis>=5.0",
    "structlog>=24.1",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "jinja2>=3.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "types-requests",
    "types-python-dateutil"
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "PL", "RUF"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
show_error_codes = true

